fix(build): correct CMake target names and int64_t casts for SQLiteCpp binding
- Qt6 minimum version lowered to 6.4 (system has 6.4.2, CI uses 6.8)
- SQLiteCpp target is 'SQLiteCpp' not 'SQLiteCpp::SQLiteCpp' on this vcpkg version
- stduuid target is 'stduuid' not 'stduuid::stduuid' on this vcpkg version
- long long is ambiguous for SQLiteCpp::Statement::bind(); use int64_t explicitly

+++ src/adapters/storage/migrations.cpp
@@ -83,7 +83,7 @@ mondoc::expected<int, mondoc::Error> runMigrations(SqliteConnection& conn) {
                 db,
                 "INSERT INTO schema_migrations(version, applied_at) VALUES(?, ?)");
             insert.bind(1, m.version);
-            insert.bind(2, static_cast<long long>(unixNowSeconds()));
+            insert.bind(2, static_cast<int64_t>(unixNowSeconds()));
             insert.exec();
             tx.commit();
             db.exec("PRAGMA user_version = " + std::to_string(m.version) + ";");
